fix: refactor + throw error
