+++ src/components/Atoms/IconNew/IconNew.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import styled, { DefaultTheme } from 'styled-components';
 
 // React-icons
-import * as icons from '../../../utils/icons';
+import * as icons from './icons';
 
 export type TIconContainerProps = {
     size?: '18' | '20' | '24' | '28' | '32';
@@ -27,12 +27,12 @@ const IconContainer = styled.svg.attrs<TIconContainerProps>(props => ({
 	viewBox: '0 0 18 18',
 	xmlns: 'http://www.w3.org/2000/svg',
 	fill: 'none',
+	height: `${props.size}px`,
+	width: `${props.size}px`,
+	strokeWidth: `${calculateStrokeWidth(props.size)}`,
 })).withConfig({
 	shouldForwardProp: (prop, defaultValidatorFn) => !['size', 'color'].includes(prop) && defaultValidatorFn(prop),
 })<TIconContainerProps>`
-	height: ${props => props.size}px;
-	width: ${props => props.size}px;
-	stroke-width: ${props => calculateStrokeWidth(props.size)};
 	color: ${props => (props.theme.color[props.color])};
   `;
 
@@ -43,8 +43,7 @@ export const IconNew: React.FC<TIconNew> = ({
 	const SvgIcon = icons[name];
 
 	if (!SvgIcon) {
-		// eslint-disable-next-line react/jsx-no-useless-fragment
-		return <></>;
+		throw new Error(`Icon "${name}" not found. Please ensure that the name is correct.`);
 	}
 
 	return (
